[update]型をFloatに変換

+++ Assets/Scripts/Runtime/1.Domain/ValueObject/Health.cs
@@ -11,7 +11,7 @@ namespace Domain
         ///     コンストラクタ
         /// </summary>
         /// <param name="hp">初期体力</param>
-        public Health(int hp)
+        public Health(float hp)
         {
             //　コンストラクタの引数が1未満の場合は例外を投げる
             //  存在してはいけないインスタンスはそもそも生成されないようにするため
@@ -22,14 +22,14 @@ namespace Domain
             _hp = hp;
         }
         /// <summary>体力のプロパティ</summary>
-        public int Hp => _hp;
+        public float Hp => _hp;
         /// <summary>
         ///     値を減らす処理
         ///     新しいHealthオブジェクトを返す
         ///     ValueObjectは値の変更ができないため、減らす処理は新しいオブジェクトを返す形で実装する
         /// </summary>
         /// <param name="damage"></param>
-        public Health TakeDamage(int damage)
+        public Health TakeDamage(float damage)
         {
             return new Health(Hp - damage);
         }
@@ -42,6 +42,6 @@ namespace Domain
         public bool Equals(Health other) => Hp == other.Hp;
 
         /// <summary>体力のフィールド変数(書き換え不能)</summary>
-        private readonly int _hp;
+        private readonly float _hp;
     }
 }
